consolidate rukpak binaries and deployments
