+++ manifests/apis/webhooks/resources/service.yaml
@@ -10,4 +10,4 @@ spec:
       protocol: TCP
       targetPort: 9443
   selector:
-    app: core-webhook
+    app: webhooks
