+++ manifests/apis/webhooks/resources/serviceaccount.yaml
@@ -1,5 +1,5 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: core-admin
+  name: webhooks-admin
   namespace: system
