+++ manifests/core/resources/bundle_reader_client_clusterrole.yaml
