+++ manifests/core/resources/bundle_uploader_client_clusterrole.yaml
