+++ manifests/core/resources/cluster_role.yaml
@@ -3,12 +3,22 @@ apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
   creationTimestamp: null
-  name: registry-provisioner-admin
+  name: core-admin
 rules:
 - nonResourceURLs:
   - /bundles/*
   verbs:
   - get
+- nonResourceURLs:
+  - /uploads/*
+  verbs:
+  - get
+- apiGroups:
+  - '*'
+  resources:
+  - '*'
+  verbs:
+  - '*'
 - apiGroups:
   - authentication.k8s.io
   resources:
@@ -36,6 +46,26 @@ rules:
   - pods/log
   verbs:
   - get
+- apiGroups:
+  - core.rukpak.io
+  resources:
+  - bundledeployments
+  verbs:
+  - list
+  - watch
+- apiGroups:
+  - core.rukpak.io
+  resources:
+  - bundledeployments/finalizers
+  verbs:
+  - update
+- apiGroups:
+  - core.rukpak.io
+  resources:
+  - bundledeployments/status
+  verbs:
+  - patch
+  - update
 - apiGroups:
   - core.rukpak.io
   resources:
