+++ manifests/core/resources/cluster_role_binding.yaml
@@ -1,13 +1,13 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: registry-provisioner-admin
+  name: core-admin
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: registry-provisioner-admin
+  name: core-admin
 subjects:
   - apiGroup: ""
     kind: ServiceAccount
-    name: registry-provisioner-admin
+    name: core-admin
     namespace: rukpak-system
