+++ manifests/core/resources/deployment.yaml
@@ -2,22 +2,22 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   namespace: rukpak-system
-  name: upload-manager
+  name: core
   labels:
-    app: upload-manager
+    app: core
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: upload-manager
+      app: core
   template:
     metadata:
       labels:
-        app: upload-manager
+        app: core
       annotations:
         kubectl.kubernetes.io/default-container: manager
     spec:
-      serviceAccountName: upload-manager-admin
+      serviceAccountName: core-admin
       containers:
         - name: kube-rbac-proxy
           image: quay.io/brancz/kube-rbac-proxy:v0.12.0
@@ -25,7 +25,7 @@ spec:
             - "--secure-listen-address=0.0.0.0:8443"
             - "--upstream=http://127.0.0.1:8080/"
             - "--logtostderr=true"
-            - "--v=1"
+            - "--v=10"
             - "--client-ca-file=/var/run/secrets/kubernetes.io/serviceaccount/ca.crt"
             - "--tls-cert-file=/etc/pki/tls/tls.crt"
             - "--tls-private-key-file=/etc/pki/tls/tls.key"
@@ -33,38 +33,36 @@ spec:
             - containerPort: 8443
               protocol: TCP
               name: https
-          securityContext:
-            allowPrivilegeEscalation: false
           volumeMounts:
             - name: certs
               mountPath: /etc/pki/tls
         - name: manager
           image: quay.io/operator-framework/rukpak:latest
           imagePullPolicy: IfNotPresent
-          command: ["/uploadmgr"]
+          command: ["/core"]
           args:
+            - "--unpack-image=quay.io/operator-framework/rukpak:latest"
+            - "--base-upload-manager-url=https://$(CORE_SERVICE_NAME).$(CORE_SERVICE_NAMESPACE).svc"
+            - "--provisioner-storage-dir=/var/cache/bundles"
+            - "--upload-storage-dir=/var/cache/uploads"
             - "--http-bind-address=127.0.0.1:8080"
-          livenessProbe:
-            httpGet:
-              path: /healthz
-              port: 8081
-            initialDelaySeconds: 15
-            periodSeconds: 20
-          readinessProbe:
-            httpGet:
-              path: /readyz
-              port: 8081
-            initialDelaySeconds: 5
-            periodSeconds: 10
-          securityContext:
-            allowPrivilegeEscalation: false
+            - "--http-external-address=https://$(CORE_SERVICE_NAME).$(CORE_SERVICE_NAMESPACE).svc"
+            - "--bundle-ca-file=/etc/pki/tls/ca.crt"
+          ports:
+            - containerPort: 8080
           volumeMounts:
             - name: bundle-cache
               mountPath: /var/cache/bundles
+            - name: upload-cache
+              mountPath: /var/cache/uploads
+            - name: certs
+              mountPath: /etc/pki/tls
       volumes:
         - name: bundle-cache
           emptyDir: {}
+        - name: upload-cache
+          emptyDir: {}
         - name: certs
           secret:
-            secretName: upload-manager-cert
+            secretName: core-cert
             optional: false
