+++ manifests/core/resources/service.yaml
@@ -2,7 +2,7 @@ apiVersion: v1
 kind: Service
 metadata:
   namespace: rukpak-system
-  name: plain-provisioner
+  name: core
 spec:
   ports:
     - name: https
@@ -10,4 +10,4 @@ spec:
       protocol: TCP
       targetPort: 8443
   selector:
-    app: plain-provisioner
+    app: core
