+++ manifests/core/resources/serviceaccount.yaml
@@ -1,5 +1,5 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: upload-manager-admin
+  name: core-admin
   namespace: rukpak-system
