--- manifests/provisioners/kustomization.yaml
@@ -1,5 +1,0 @@
-bases:
-- ./common
-- ./plain
-- ./registry
-
