--- manifests/provisioners/plain/resources/certificate.yaml
@@ -1,14 +1,0 @@
----
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: plain-provisioner
-  namespace: rukpak-system
-spec:
-  secretName: plain-provisioner-cert
-  dnsNames:
-    - $(PLAIN_PROVISIONER_SERVICE_NAME).$(PLAIN_PROVISIONER_SERVICE_NAMESPACE).svc
-    - $(PLAIN_PROVISIONER_SERVICE_NAME).$(PLAIN_PROVISIONER_SERVICE_NAMESPACE).svc.cluster.local
-  issuerRef:
-    kind: Issuer
-    name: rukpak-ca-issuer
