--- manifests/provisioners/plain/resources/cluster_role.yaml
@@ -1,86 +1,0 @@
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  creationTimestamp: null
-  name: plain-provisioner-admin
-rules:
-- nonResourceURLs:
-  - /bundles/*
-  verbs:
-  - get
-- apiGroups:
-  - '*'
-  resources:
-  - '*'
-  verbs:
-  - '*'
-- apiGroups:
-  - authentication.k8s.io
-  resources:
-  - tokenreviews
-  verbs:
-  - create
-- apiGroups:
-  - authorization.k8s.io
-  resources:
-  - subjectaccessreviews
-  verbs:
-  - create
-- apiGroups:
-  - ""
-  resources:
-  - pods
-  verbs:
-  - create
-  - delete
-  - list
-  - watch
-- apiGroups:
-  - ""
-  resources:
-  - pods/log
-  verbs:
-  - get
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundledeployments
-  verbs:
-  - list
-  - watch
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundledeployments/finalizers
-  verbs:
-  - update
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundledeployments/status
-  verbs:
-  - patch
-  - update
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundles
-  verbs:
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundles/finalizers
-  verbs:
-  - update
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundles/status
-  verbs:
-  - patch
-  - update
