--- manifests/provisioners/plain/resources/serviceaccount.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: plain-provisioner-admin
-  namespace: rukpak-system
