--- manifests/provisioners/registry/kustomization.yaml
@@ -1,21 +1,0 @@
-resources:
-  - resources/certificate.yaml
-  - resources/cluster_role.yaml
-  - resources/cluster_role_binding.yaml
-  - resources/deployment.yaml
-  - resources/service.yaml
-  - resources/serviceaccount.yaml
-
-vars:
-  - name: REGISTRY_PROVISIONER_SERVICE_NAMESPACE # namespace of the service
-    objref:
-      kind: Service
-      version: v1
-      name: registry-provisioner
-    fieldref:
-      fieldpath: metadata.namespace
-  - name: REGISTRY_PROVISIONER_SERVICE_NAME
-    objref:
-      kind: Service
-      version: v1
-      name: registry-provisioner
