--- manifests/provisioners/registry/resources/certificate.yaml
@@ -1,14 +1,0 @@
----
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: registry-provisioner
-  namespace: rukpak-system
-spec:
-  secretName: registry-provisioner-cert
-  dnsNames:
-    - $(REGISTRY_PROVISIONER_SERVICE_NAME).$(REGISTRY_PROVISIONER_SERVICE_NAMESPACE).svc
-    - $(REGISTRY_PROVISIONER_SERVICE_NAME).$(REGISTRY_PROVISIONER_SERVICE_NAMESPACE).svc.cluster.local
-  issuerRef:
-    kind: Issuer
-    name: rukpak-ca-issuer
