--- manifests/provisioners/registry/resources/serviceaccount.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: registry-provisioner-admin
-  namespace: rukpak-system
