--- manifests/uploadmgr/resources/certificate.yaml
@@ -1,18 +1,0 @@
----
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: upload-manager
-  namespace: rukpak-system
-spec:
-  secretName: upload-manager-cert
-  dnsNames:
-    - $(UPLOAD_MANAGER_SERVICE_NAME).$(UPLOAD_MANAGER_SERVICE_NAMESPACE).svc
-    - $(UPLOAD_MANAGER_SERVICE_NAME).$(UPLOAD_MANAGER_SERVICE_NAMESPACE).svc.cluster.local
-
-    # Include localhost to support rukpakctl run (which uses port-forwarding to expose this service on "localhost")
-    # This is required to allow rukpakctl's certificate verification to succeed.
-    - localhost
-  issuerRef:
-    kind: Issuer
-    name: rukpak-ca-issuer
