--- manifests/uploadmgr/resources/cluster_role.yaml
@@ -1,33 +1,0 @@
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: upload-manager-admin
-rules:
-- apiGroups:
-  - authentication.k8s.io
-  resources:
-  - tokenreviews
-  verbs:
-  - create
-- apiGroups:
-  - authorization.k8s.io
-  resources:
-  - subjectaccessreviews
-  verbs:
-  - create
-- apiGroups:
-  - core.rukpak.io
-  resources:
-  - bundles
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
-    - core.rukpak.io
-  resources:
-    - bundles/status
-  verbs:
-    - update
-    - patch
