--- manifests/uploadmgr/resources/cluster_role_binding.yaml
@@ -1,13 +1,0 @@
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: upload-manager-admin
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: upload-manager-admin
-subjects:
-  - apiGroup: ""
-    kind: ServiceAccount
-    name: upload-manager-admin
-    namespace: rukpak-system
