--- manifests/uploadmgr/resources/service.yaml
@@ -1,13 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  namespace: rukpak-system
-  name: upload-manager
-spec:
-  ports:
-    - name: https
-      port: 443
-      protocol: TCP
-      targetPort: 8443
-  selector:
-    app: upload-manager
